Updated README and Manual. Removed Manual PDF.
Removed Manual PDF because the Markdown version is so much better, and
easier to update. The PDF manual gets outdated easily, and needs to be
generated from the MD, which is tedious and tends to generate errors.

README and Manual were updated to include the rationale for the proof
checker tool.